requirements.txt
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.15
cryptography==46.0.4
filelock==3.20.3
Flask==3.1.2
fsspec==2026.1.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
PyAudio==0.2.14
pycparser==3.0
pypdfium2==5.3.0
pypiwin32==223
python-docx==1.2.0
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shellingham==1.5.4
SpeechRecognition==3.14.5
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
